update from inliniac/suricata

+++ src/detect-engine-hsbd.c
@@ -190,7 +190,7 @@ static const uint8_t *DetectEngineHSBDGetBufferForTX(htp_tx_t *tx, uint64_t tx_i
     if (htud->response_body.body_inspected > htp_state->cfg->response.inspect_min_size) {
         BUG_ON(htud->response_body.content_len_so_far < htud->response_body.body_inspected);
         uint64_t inspect_win = htud->response_body.content_len_so_far - htud->response_body.body_inspected;
-        SCLogDebug("inspect_win %u", (uint)inspect_win);
+        SCLogDebug("inspect_win %"PRIu64, inspect_win);
         if (inspect_win < htp_state->cfg->response.inspect_window) {
             uint64_t inspect_short = htp_state->cfg->response.inspect_window - inspect_win;
             if (htud->response_body.body_inspected < inspect_short)
